backend/requirements.txt
# Web framework
Flask==3.0.0

# Database (PostgreSQL) - updated for Python 3.13 compatibility
SQLAlchemy>=2.0.30
psycopg[binary]>=3.1.0
alembic>=1.13.0

# Google Cloud / Firestore
google-cloud-firestore==2.14.0

# Environment / config
python-dotenv==1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0
